Change workers for reports build
